requirements.txt
fastapi
uvicorn
httpx
pydantic
google-adk
google-cloud-aiplatform
litellm
streamlit
python-dotenv
aiofiles
